Updated Readme with a Fork section
When I started out to help update the site I spent hours trying to figure this out. For developers I'm sure it takes only a few minutes. My point of view is that sometimes the best documentation is by someone who is not involved with a system's development. I think these updates proposed by me make it much easier for a non-developer to help maintain the website without moving over to something like Wordpress.